IO ROM
This is the data for the 2716 rom chip. This controls what and where IO devices show up on the address space. For example, if you wanted device "4" to be on port "240" (0xF0) of the computers address space then you would put the number "04" at address "240" (0xF0) on the rom chip.